+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Change Log
 
-## [master](https://github.com/Vierbeuter/craft-footnotes/tree/master) (2017-\*-\*)
-Minor improvement: Set icon for footnote button.
+## Unreleased
+
+## [2.0.0](https://github.com/Vierbeuter/craft-footnotes/tree/2.0.0) (2019-10-29)
+
+### Changed
+ - update sources for Craft CMS 3
+
+### Added
+ - minor improvement: Set icon for footnote button
 
 ## [1.1.3](https://github.com/Vierbeuter/craft-footnotes/tree/1.1.3) (2017-05-10)
 Minor improvement: Footnotes directly followed by other footnotes (either with or without whitespace characters between them) will now be comma-separated: